Updated response types
